pak finder improvements/fixes
Don't pop a dialog box if checking a CD/DVD drive that was recently used (but is now empty). Addresses issue #15.

Also fix the for loop for iterating over the drives when scanning common paths. Just didn't work! ... was only checking the first drive in the list. Batch file for-syntax makes me something something.