fix soldier idle sound (in some engines)
In QSS, loose files can override pak contents. Might be true of some other engines also?

Anyway, this idle.wav file in the warpspasm pak has no samples in it (because it's supposed to be just silence), which causes QSS to spam console warnings each time it tries to play the file. We'll replace it here with a file that contains a small number of silent samples.