store Quake engine and launch args in "_last_launch.cmd" in mod folder
Currently this is just informative. But in future Quakestarter versions, if a new version of AD or Copper has come out, we can use this info to deal with currently installed mods that might have been running against the older version. I'm not 100% sure what to do in that case, but the "legacies" system does give us some options.