script to help find qualifying releases
I don't think the Quaddicted database.xml used to expose the normalized ratings, but now it does... cool! Easier to automate finding the releases for the installer menus.

It's still kind of a manual process to find which things actually qualify as an episode/hub (i.e. have a real startmap + minimum number of maps). Taking a best-guess at it here, but some of the things in the output "episodes" list really should be evaluated as "other" (for example czg07).

Trying out a ratings scheme that is a bit more generous, since the Quaddicted normalized ratings seem to have taken a downward trend overall. Haven't gone through this in detail yet though to see what these new criteria would result in adding/removing.