pak hash check and expanded music dir check
Support looking in rerelease dirs in order to find music dirs there.

ALSO do pak file hash checking, so we don't grab the rerelease paks as opposed to the initial paks. This also helps us avoid grabbing the 1.01 pak0.pak.

Might want to add config option to allow intentionally looking for the rerelease paks? What complicates things a bit there is that the rerelease doesn't have pak1.pak, and there are definitely some spots in these scripts that assume the necessity of pak1.pak.

This would address issue #43 and make issue #69 a moot point. It does most of issue #63; next we just need to remove the soundtrack-downloading code. And update the docs.